Changed code to be more objecty-oriented-y.
By which I mean that datatypes are now classes, tags are now classes
(with a factory that produces new ones based on the tag type you give it),
and when I get around to doing them ActionRecords will be classes too.

Well, okay, not quite all datatypes are classes - some are just lists and
have ended up as lists, and the string datatype still has datatypes.string()
as the method to sort it out, but just returns a Python 2.x string now.