Added Swf.included_tags() for extra info about tags
Running swf.py now gets you, at the end of the info about each tag, a list
of all the tags in that file sorted by frequency (does not include tags
embedded in DefineSprite*), and a note about whether they have parsers
implemented yet. This helps with deciding which parsers to implement next.