Added DefineFont (mostly) and a lot of datatypes
Added a parser for most of DefineFont, and many new datatypes that
are required for it and its attendant datatypes. Either Shape and
Shaperecords, or DefineFont itself, aren't working properly; with
one test file, parsing the glyph shapes in a DefineFont tag runs
off the end of the tag stream. Have commented the glyph shape parsing
out for now; will come back to it later.

Also removed all of the old-style datatypes that have been implemented
as classes. There's only one left now.