Trying to fix DefineShape issues, not yet successful
Something in the ShapeWithStyle parsing is going wrong; two of my test files
are having problems with parsing depending on whether one bit of a file
is forced to be byte-aligned or not. Will look into in more detail, but for
now am adding remaining changes before putting on Github.